pyproject.toml
[tool.poetry]
name = "charm-json"
version = "0.1.0"
description = ""
authors = ["Carl Csaposs <carl.csaposs@canonical.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
charm-api = {git = "https://github.com/canonical/charm-api"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
